refactor: split renderer diagnostics implementation
